+++ CHANGELOG.md
@@ -5,6 +5,8 @@
   - Use more optimal SQL when polling workflows when database supports update returning syntax
   - Only rollback poll operation when no workflows could be allocated for executing (when multiple pollers compete for same workflows)
   - Allow configuring executor queue length with _nflow.dispatcher.executor.queue.size_
+- nflow-rest:
+  - Add support for user-provided action description when updating a workflow instance
 
 ## 1.2.0 (2014-12-23)